python 2

+++ Personal/calculator.py
@@ -2,7 +2,8 @@
 num1 = input("Enter a number: ")
 num2 = input("Enter a second number: ")
 
-# parsing the string to the number
-result = int(num1) + int(num2)
+# parsing the string to the number, if its decimal it will be error out
+#result = int(num1) + int(num2)
+result = float(num1) + float(num2)
 
 print(result)
